Move 'check_maven_artifact_map' to task.py
