+++ beetmoverscript/tests/test_script.py
@@ -13,7 +13,6 @@ import beetmoverscript.script
 from beetmoverscript.constants import PARTNER_REPACK_PRIVATE_REGEXES, PARTNER_REPACK_PUBLIC_REGEXES
 from beetmoverscript.script import (
     async_main,
-    check_maven_artifact_map,
     copy_beets,
     enrich_balrog_manifest,
     get_destination_for_partner_repack_path,
@@ -68,45 +67,6 @@ async def test_push_to_releases(context, mocker, candidates_keys, releases_keys,
             await push_to_releases(context)
     else:
         await push_to_releases(context)
-
-
-def test_check_maven_artifact_map(context):
-    context.action = "push-to-maven"
-
-    fake_correct_version = "12.3.20200920201111"
-    source = "fake_path/fake-artifact-{version}.jar"
-
-    def artifact_map_entry_with_version(version):
-        return {
-            "locale": "en-US",
-            "paths": {
-                source: {
-                    "checksums_path": "",
-                    "destinations": [f"fake/destination/{version}/fake-artifact-{version}.jar"],
-                }
-            },
-            "taskId": "fake-task-id",
-        }
-
-    context.task = {
-        "payload": {
-            "artifactMap": [],
-            "releaseProperties": {"appName": "nightly_components"},
-            "upstreamArtifacts": [{"paths": [source], "taskId": "fake-task-id", "taskType": "build"}],
-            "version": "bad.version",
-        }
-    }
-
-    with pytest.raises(ScriptWorkerTaskException):
-        check_maven_artifact_map(context)
-
-    context.task["payload"]["version"] = fake_correct_version
-    context.task["payload"]["artifactMap"] = [artifact_map_entry_with_version(fake_correct_version)]
-    check_maven_artifact_map(context)
-
-    context.task["payload"]["artifactMap"] = [artifact_map_entry_with_version("a.bad.version")]
-    with pytest.raises(ScriptWorkerTaskException):
-        check_maven_artifact_map(context)
 
 
 # copy_beets {{{1
